+++ src/pyejabberd/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.2.3"
+__version__ = "0.2.4"
 
 try:  # pragma: no cover
     # https://urllib3.readthedocs.org/en/latest/security.html#pyopenssl