+++ src/pyejabberd/defaults.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals, print_function
+
+
+XMLRPC_API_PROTOCOL = 'https'
+XMLRPC_API_PORT = 4560